My project in FPGA programing
It is my first project, which i did during logic deisgn course in the university.
In this project I first created my code on ModelSim with verilog and checked it over some test benches. I then passed the code for a hardware check in Quartus and finally I programed it to the DE2-115 board which contains FPGA device.